Allow packing of test executions.
In case of e.g., large test suites with many tests startup time of the
test executables can contribute significantly to the test execution
time.

Here we add an additional parameter `-p`/`--pack` which if given causes
tests to be executed in sets of `option.workers`. This minimizes the
amount of forked processes while trying to maximize parallelism.

We examining the known test times and group test in such a way that each
group has roughly similar summed timings. Tests with unknown execution
time are always executed in isolation to obtain a measurement. Note that
these timing are only approximate proxies for the time it takes to
execute the test in a larger set as they include e.g., the overhead of
forking the process or the runtime of global or test case fixture setup
or teardown.